ints should be stored as bigint. cost is small and makes dealing with large numbers like fragment ids much easier

+++ src/funlib/persistence/graphs/pgsql_graph_database.py
@@ -205,7 +205,7 @@ class PgSQLGraphDatabase(SQLGraphDataBase):
         if isinstance(type, Vec):
             return self.__sql_type(type.dtype) + f"[{type.size}]"
         try:
-            return {bool: "BOOLEAN", int: "INTEGER", str: "VARCHAR", float: "REAL"}[
+            return {bool: "BOOLEAN", int: "BIGINT", str: "VARCHAR", float: "REAL"}[
                 type
             ]
         except ValueError:
